chroniccare-backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
psycopg[binary]==3.2.4
asyncpg==0.30.0
alembic==1.13.3

# Redis
redis==5.1.1

# Settings
pydantic-settings==2.5.2
pydantic[email]==2.12.5

# AI / LLM
openai==1.51.0

# Utils
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
httpx==0.27.2

greenlet>=3.0.0
email-validator>=2.0.0
